Small changes to the benchmarking. Changed the number of searches.
git-svn-id: http://svn.seqan.de/seqan/trunk@10911 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/sandbox/singer/include/seqan/fm_sequence/bit_string.h
@@ -72,7 +72,7 @@ struct Fibre<RankSupportBitString<TSpec>, FibreRankSupportBucketString>
 template <typename TSpec>
 struct Fibre<RankSupportBitString<TSpec>, FibreRankSupportSuperBucketString>
 {
-    typedef String<unsigned long> Type;
+    typedef String<unsigned int> Type;
 };
 
 template <typename TSpec>
